Updated range sensor API more consistent with others. Added canned nav data to test SPI interface more accurately. Added new conditional compilation flags to make testing more structured

+++ libraries/Ultrasonic/Ultrasonic.h
@@ -43,7 +43,7 @@ class Ultrasonic : public Sensor
   public:
     Ultrasonic(int Trigger, int Echo);
     ~Ultrasonic();
-    unsigned short ReadRange();				// Read the range data
+    int ReadRange();						// Read the range data
 
     // Accessor
     unsigned short	getRange();
